makefile: make runc non-PHONY target
Making runc a .PHONY target means that we rebuild it each time, even if
the source code does not change. This can cause issues if you are trying
attach uprobes a binary that then gets replaced when you try to run
integration tests with it.

The "static" target needed to be adjusted a little, but now it's just a
special case of the standard "runc" target.

Also, the "runc-bin" and "static-bin" rules were really not necessary in
the first place...

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>