libct/nsenter: better read/write errors
Introduce and use iobail, xread, and xwrite wrappers so that we can
properly check read/write return value and call either bail or bailx on
error, with proper diagnostics (distinguishing failed read/write from a
short read/write).

This prevents the "Success" prefix in errors like:

	failed to sync with stage-1: next state: Success

This also fixes a subtle bug of not calling sane_kill in one case.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>