libct: switch final WithProcfd users to WithProcfdFile
This probably should've been done as part of commit d40b3439a961
("rootfs: switch to fd-based handling of mountpoint targets") but it
seems I missed them when doing the rest of the conversions.

This also lets us remove utils.WithProcfd entirely, as well as
pathrs.MkdirAllInRoot. Unfortunately, WithProcfd was exposed in the
externally-importable "libcontainer/utils" package and so we need to
have a deprecation notice to remove it in runc 1.5.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
(cherry picked from commit 9dbd37e06f8124be6e496308e1735d18f27bb730)
Signed-off-by: Aleksa Sarai <aleksa@amutable.com>